Done some changes and additions to database
Modified a bit Requests table and added a new one called Responses to handle the invitations for committee members. Created two new triggers, one to insert a record in Responses for each new record in Requests and another one that handles the responses on the invitations and updates Thesis and Requests tables accordingly. Also, created a new procedure which is called from the handle_invitation.php to update the Responses table because of some is with the direct query.