Fixed a bug and added a feature
Fixed a bugged that was updates the status of a request for member committee to 'Declined' from 'Accepted' when the committee was filled and the thesis status updated to 'Active'. Also added two buttons that will allow a professor to add notes to an 'Active' thesis he is a member of or he supervises. (buttons have no functionality yet)